create a specialized `type_caster` template for `std::shared_ptr<ov::Model>`
